Update and rename sgraph_v1.py to sgraph_v2.py
Version 1:
Consisted of one python script utilizing the OpenCV library to access camera and output live video feed in the main window. This code could also take screenshots and dynamically name them, as well as store them in a file. These functions relied on inputs from a keyborad.

Version 2:
Utilizing the Tkinter and Pillow libraries in addition to OpenCV, Version 2 integrates a GUI to the video feed with buttons able to execute the functions of Version 1. This code no longer relies on keyboard inputs and instead utilizes button presses; the buttons themselves are overlaid on top of the live feed.